fundextract/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pdfplumber==0.11.0
camelot-py[cv]==0.11.0
anthropic==0.28.0
rapidfuzz==3.9.3
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
Pillow==10.3.0
pandas==2.2.2
numpy==1.26.4
